Cargo.toml: added new feature `cc_kbc_occlum`
`cc_kbc_occlum` enables occlum attester and cc-kbc in attestation-agent
dependency, which can be used in enclave-cc to support occlum attester.

CI also updated to test this new feature.

Also, because the refactoring of AA, we need to introduce two different
crate `kbc` and `attestation_agent` to work the same as before.

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>